refactor: alterar estrutura de pastas
