+++ src/core/errors/bad-request.error.ts
