+++ src/core/errors/conflict.error.ts
