+++ src/core/errors/internal-server.error.ts
