+++ src/core/models/irrigation.model.ts
