+++ src/core/models/pivot.model.ts
