+++ src/routers/auth.routes.ts
@@ -1,9 +1,9 @@
 import { Router } from 'express';
-import * as authController from './auth.controller';
-import { withErrorHandler } from '../../handlers/with-error-handler';
-import { withValidationHandler } from '../../handlers/with-validation-handler';
-import { registerUserRequestSchema } from './schemas/register-user-request.schema';
-import { loginRequestSchema } from './schemas/login-request.schema';
+import * as authController from '../controllers/auth.controller';
+import { withErrorHandler } from '../middlewares/with-error-handler';
+import { withValidationHandler } from '../middlewares/with-validation-handler';
+import { registerUserRequestSchema } from '../schemas/register-user-request.schema';
+import { loginRequestSchema } from '../schemas/login-request.schema';
 
 export function createAuthRouter(): Router {
   const router = Router();
