+++ src/routers/irrigation.router.ts
@@ -1,10 +1,10 @@
 import { Router } from 'express';
-import { verifyJwt } from '../../middlewares/verify-jwt.middleware';
-import { withValidationHandler } from '../../handlers/with-validation-handler';
-import { irrigationRequestSchema } from './schemas/irrigation-request.schema';
-import { withErrorHandler } from '../../handlers/with-error-handler';
+import { verifyJwt } from '../middlewares/verify-jwt';
+import { withValidationHandler } from '../middlewares/with-validation-handler';
+import { irrigationRequestSchema } from '../schemas/irrigation-request.schema';
+import { withErrorHandler } from '../middlewares/with-error-handler';
 
-import * as irrigationController from './irrigation.controller';
+import * as irrigationController from '../controllers/irrigation.controller';
 
 export function createIrrigationRouter(): Router {
   const router = Router();
