+++ src/schemas/login-request.dto.ts
@@ -1,4 +1,4 @@
 import z from "zod";
-import { loginRequestSchema } from "../schemas/login-request.schema";
+import { loginRequestSchema } from "./login-request.schema";
 
 export type LoginRequest = z.infer<typeof loginRequestSchema>;
