+++ src/schemas/pivot-request.schema.ts
