+++ src/schemas/register-user-request.dto.ts
@@ -1,4 +1,4 @@
 import z from "zod";
-import { registerUserRequestSchema } from "../schemas/register-user-request.schema";
+import { registerUserRequestSchema } from "./register-user-request.schema";
 
 export type RegisterUserRequest = z.infer<typeof registerUserRequestSchema>;
