+++ src/schemas/register-user-request.schema.ts
