+++ src/schemas/register-user-response.dto.ts
